remove cred
